Remove debugging, stop spatial only if it started

+++ ci/run-tests.ps1
@@ -85,9 +85,8 @@ if($run_with_spatial) {
 Echo "Running $($ue_path_absolute) $($cmd_args_list)"
 
 $run_tests_proc = Start-Process $ue_path_absolute -PassThru -NoNewWindow -ArgumentList $cmd_args_list
-Start-Sleep -s 10
-$tail_proc = Start-Process "powershell" -PassThru -ArgumentList @("Get-Content", $log_file_path, "-Wait")
 Wait-Process -Id (Get-Process -InputObject $run_tests_proc).id
-Stop-Process -InputObject $tail_proc
 
-Stop-Process -InputObject $spatial_process
+if ($run_with_spatial) {
+    Stop-Process -InputObject $spatial_process
+}
